Rename ModelCompasionProsesipynb to ModelCompasionProcessipynb
This program is a process for carrying out model comparisons with results, sample images must be trained one by one into the comparative model. Adjust the model name and data, pay attention to the location where the model and data are stored so that the training can run as it should and produce results.